requirements.txt
librosa
pytorch==1.10.0
scipy